Add files via upload
get-me-some.txt is a text file created by me that contains sentences of what would probably be orders of my friends from shops as well as thier occupation. All I did was to pre-process these sentences using python and organizing it in a pandas Dataframe. I also counted the words in the text file just for fun and ploted it on a graph. The code and analysis can be found in the Text-to-Dataframe notebook.